finito e funzionante più o meno
Creata classe per gestire i parametri del json, creato metodo per ottenere  la lista di tutti i messaggi, implementata la metodologia con offset, corretti errori nel Thread.
Manca solo la scrittura sul csv

+++ TelegramBot/src/telegrambot/OpenStreetMap.java
@@ -29,7 +29,7 @@ public class OpenStreetMap {
 
        
     
-public void rispondi(String citta,int idDestinatario)
+public void rispondi(String citta,long idDestinatario)
 {
         BufferedReader in = null;
         PrintWriter out;
